Switched long to int

+++ src/main/kotlin/io/cyborgsquirrel/home/responses/HomeResponse.kt
@@ -4,8 +4,8 @@ import io.micronaut.serde.annotation.Serdeable
 
 @Serdeable
 data class HomeResponse(
-    val clients: Long,
-    val strips: Long,
-    val effects: Long,
-    val palettes: Long,
+    val clients: Int,
+    val strips: Int,
+    val effects: Int,
+    val palettes: Int,
 )
